requirements.txt
asgiref==3.8.1
Django==4.2
psycopg2-binary==2.9.7
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
